Updated get_effwavelens to also grab throughputs version info.
