+++ python/lsst/sims/seeingModel/__init__.py
@@ -1,6 +1,7 @@
 from .seeingData import *
 from .seeingModel import *
 from .seeingModelConfig import *
+from .utils import *
 from .version import *
